+++ src/components/Borrow/BorrowForm.tsx
@@ -1,0 +1,134 @@
+import React, { useContext } from "react";
+
+import FirstStageTable from "./FirstStageTable";
+import SecondStageTable from "./SecondStageTable";
+import ThirdStageTable from "./ThirdStageTable";
+import PrimaryButton from "../UI/PrimaryButton";
+import LoginButton from "../LoginButton/LoginButton";
+
+import "./borrow.scss";
+
+import {
+  AppContext,
+  AppContextState,
+  TellerTokens,
+  BaseTokens,
+} from "../../context/app";
+
+import {
+  BorrowPageContext,
+  LendingApplicationMap,
+} from "../../context/borrowContext";
+
+import { sendLendingApplication } from "../../models/ArrowheadCRA";
+import { getLendingPoolDecimals } from "../../models/Contracts";
+import { getNonce } from "../../models/DataProviders";
+
+const BorrowForm = () => {
+  const {
+    stage,
+    setStage,
+    submenu,
+    borrowRequest,
+    borrowProcessState,
+  } = useContext(BorrowPageContext);
+  const { state, updateAppState } = useContext(AppContext);
+  const { setRequesting, setSuccess, setSubmitting } = borrowProcessState;
+
+  const requestLoan = async () => {
+    const { dataProviderResponse, web3State } = state;
+    //TODO: this should update based on the selected ATM type.
+    const { lendingPool } = state.teller.contracts[BaseTokens.ETH][
+      TellerTokens.tDAI
+    ];
+    try {
+      const tokenDecimals = await getLendingPoolDecimals(
+        lendingPool,
+        web3State
+      );
+      const nonceDataResponse = await getNonce();
+      const lendingApplication = LendingApplicationMap(
+        borrowRequest,
+        dataProviderResponse.bankInfo,
+        nonceDataResponse,
+        tokenDecimals,
+        web3State
+      );
+      const response = await sendLendingApplication(lendingApplication);
+      console.log(response.data);
+      return true;
+    } catch (err) {
+      console.log(err);
+      updateAppState((st: AppContextState) => {
+        const errorModal = {
+          show: true,
+          message:
+            "An error occurred during the borrowing process. Please try again.",
+          title: "Error",
+        };
+        return { ...st, errorModal };
+      });
+      return false;
+    }
+  };
+
+  const loggedIn = state.web3State?.address || "";
+  const onRequestLoan = async () => {
+    setRequesting(true);
+    const res = await requestLoan();
+    setRequesting(false);
+    setSuccess(res);
+  };
+  const onAcceptTerms = async () => {
+    setSubmitting(true);
+    //Accept loan terms
+    await new Promise((resolve) => setTimeout(resolve, 2000));
+
+    setSubmitting(false);
+    setStage(stage + 1);
+  };
+
+  return (
+    <div>
+      {submenu ? (
+        submenu
+      ) : (
+        <div>
+          {stage === 1 && (
+            <div>
+              <FirstStageTable />
+              {loggedIn ? (
+                <PrimaryButton
+                  text="Get Loan Terms"
+                  onClick={() => {
+                    //Get LoanTerms
+                    setStage(stage + 1);
+                  }}
+                />
+              ) : (
+                <LoginButton />
+              )}
+            </div>
+          )}
+          {stage === 2 && (
+            <div>
+              <SecondStageTable />
+              <PrimaryButton text="Accept Loan Terms" onClick={onAcceptTerms} />
+            </div>
+          )}
+          {stage === 3 && (
+            <div>
+              <ThirdStageTable />
+              <PrimaryButton
+                disabled={!borrowRequest.transferred}
+                text="Request Loan"
+                onClick={onRequestLoan}
+              />
+            </div>
+          )}
+        </div>
+      )}
+    </div>
+  );
+};
+export default BorrowForm;
