+++ src/components/Borrow/StageBar.tsx
@@ -28,7 +28,7 @@ const StageBar = () => {
   const { state } = useContext(AppContext);
 
   const [progress, setProgress] = useState(0);
-  const { stage, setStage } = useContext(BorrowPageContext);
+  const { stage, setStage, borrowProcessState } = useContext(BorrowPageContext);
   const loggedIn = state.web3State?.address || "";
 
   useEffect(() => {
@@ -51,12 +51,21 @@ const StageBar = () => {
       <div className={`d-flex flex-row justify-content-between`}>
         {stages.map((s) => {
           return (
-            <div key={s} className={` ${loggedIn ? "pointer" : "disabled"}`}>
+            <div
+              key={s}
+              className={` ${
+                loggedIn && s === stage - 1 ? "pointer" : "disabled"
+              }`}
+            >
               <StageNumber
                 number={s}
                 onClickAction={() => {
-                  if (loggedIn) {
-                    setStage(s);
+                  if (loggedIn && s === stage - 1) {
+                    if (stage === 3) {
+                      borrowProcessState.setStageChangeWarning(true);
+                    } else {
+                      setStage(s);
+                    }
                   }
                 }}
                 stage={stage}
